app/requirements.txt
Flask==3.0.3
gunicorn==21.2.0
mysql-connector-python==9.0.0
